Dedicated controllers for almost every API call
