+++ controllers/DeactivateAccountController.js
@@ -1,0 +1,12 @@
+const ActionLogger = require("../util/ActionLogger");
+
+exports.postAPIDeactivate = (req, res, next) => {
+    if (!req.body.password) return res.status(400).json({sucess: false, error: {message: "The password field is required.", code: "invalid_parameters"}});
+    req.user.comparePassword(req.body.password, (error, match) => {
+        if(!match || error) return res.status(401).json({success: false, error: {message: "The password you entered was incorrect.", code: "incorrect_password"}});
+        req.user.deactivated = true;
+        ActionLogger.log("deactivate", req.user);
+        req.user.save();
+        return res.json({success: true});
+    });
+}
